Rename LnNode to InvoiceHandler and move to lib
